+++ R/read_emigration.R
@@ -17,14 +17,17 @@
 #' @family Microdata
 #' @examplesIf identical(tolower(Sys.getenv("NOT_CRAN")), "true")
 #' # return data as arrow Dataset
-#' df <- read_emigration(year = 2010,
-#'                       showProgress = FALSE)
-#'
+#' df <- read_emigration(
+#'   year = 2010,
+#'   showProgress = FALSE
+#'   )
 #'
 #' # return data as data.frame
-#' df <- read_emigration(year = 2010,
-#'                       as_data_frame = TRUE,
-#'                       showProgress = FALSE)
+#' df <- read_emigration(
+#'   year = 2010,
+#'   as_data_frame = TRUE,
+#'   showProgress = FALSE
+#'   )
 #'
 #'
 read_emigration <- function(year = 2010,