+++ R/read_households.R
@@ -19,11 +19,13 @@
 #' @family Microdata
 #' @examplesIf identical(tolower(Sys.getenv("NOT_CRAN")), "true")
 #' # return data as arrow Dataset
-#' df <- read_households(year = 2010,
-#'                       showProgress = FALSE)
+#' df <- read_households(
+#'   year = 2010,
+#'   showProgress = FALSE
+#'   )
 #'
 #'
-read_households <- function(year = 2010,
+read_households <- function(year,
                             columns = NULL,
                             add_labels = NULL,
                             as_data_frame = FALSE,
@@ -31,7 +33,7 @@ read_households <- function(year = 2010,
                             cache = TRUE){
 
   ### check inputs
-  checkmate::assert_numeric(year)
+  checkmate::assert_numeric(year, any.missing = FALSE)
   checkmate::assert_vector(columns, null.ok = TRUE)
   checkmate::assert_logical(as_data_frame)
   checkmate::assert_string(add_labels, pattern = 'pt', null.ok = TRUE)